# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Netwrk API Documentation",
        "description": "API documentation for Netwrk platform - User registration and management",
        "contact": {
            "email": "support@netwrk.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local development server"
        },
        {
            "url": ""
        }
    ],
    "paths": {
        "/admins/dashboard/stats": {
            "get": {
                "tags": [
                    "admin"
                ],
                "summary": "Get admin dashboard statistics",
                "operationId": "a6b1d889051df39808b2b7f995bb7cb2",
                "responses": {
                    "200": {
                        "description": "Dashboard statistics retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Dashboard statistics retrieved successfully"
                                        },
                                        "data": {
                                            "properties": {
                                                "totalUsers": {
                                                    "type": "integer",
                                                    "example": 1250
                                                },
                                                "totalContacts": {
                                                    "type": "integer",
                                                    "example": 15600
                                                },
                                                "totalGroups": {
                                                    "type": "integer",
                                                    "example": 340
                                                },
                                                "activeUsers": {
                                                    "type": "integer",
                                                    "example": 1180
                                                },
                                                "newUsersThisMonth": {
                                                    "type": "integer",
                                                    "example": 45
                                                },
                                                "totalReferrals": {
                                                    "type": "integer",
                                                    "example": 890
                                                },
                                                "platformGrowth": {
                                                    "properties": {
                                                        "users": {
                                                            "type": "integer",
                                                            "example": 12
                                                        },
                                                        "contacts": {
                                                            "type": "integer",
                                                            "example": 8
                                                        },
                                                        "groups": {
                                                            "type": "integer",
                                                            "example": 15
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 401
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admins/dashboard/topUsers": {
            "get": {
                "tags": [
                    "admin"
                ],
                "summary": "Get top users for admin dashboard",
                "operationId": "1c3f6f98ef4a05e10e944fc2040bb247",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of top users to return",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    },
                    {
                        "name": "period",
                        "in": "query",
                        "description": "Time period for ranking (week, month, year)",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "default": "month",
                            "enum": [
                                "week",
                                "month",
                                "year"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Top users retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Top users retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "uuid"
                                                    },
                                                    "firstName": {
                                                        "type": "string",
                                                        "example": "John"
                                                    },
                                                    "lastName": {
                                                        "type": "string",
                                                        "example": "Doe"
                                                    },
                                                    "email": {
                                                        "type": "string",
                                                        "example": "john@example.com"
                                                    },
                                                    "avatar": {
                                                        "type": "string",
                                                        "example": "https://example.com/avatar.jpg"
                                                    },
                                                    "contactsCount": {
                                                        "type": "integer",
                                                        "example": 150
                                                    },
                                                    "groupsCount": {
                                                        "type": "integer",
                                                        "example": 12
                                                    },
                                                    "referralsCount": {
                                                        "type": "integer",
                                                        "example": 8
                                                    },
                                                    "lastActive": {
                                                        "type": "string",
                                                        "example": "2024-01-15T10:30:00Z"
                                                    },
                                                    "joinedAt": {
                                                        "type": "string",
                                                        "example": "2024-01-01T00:00:00Z"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 401
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthorized"
                                        },
                                        "data": {
                                            "type": "null"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/admins/dashboard/topGroups": {
            "get": {
                "tags": [
                    "admin"
                ],
                "summary": "Get top groups for admin dashboard",
                "operationId": "70e9f854233497e82eebf2c5b5be4f0d",
                "parameters": [
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Number of top groups to return",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 10
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Top groups retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "statusCode": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Top groups retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}